Enable ATen in Caffe2 (#26)
* Enable ATen in Caffe2

* Try printing stacktrace when there are segfaults in tests

* Fix bash syntax error

* Disable ccache temporarily

* Revert "Disable ccache temporarily"

This reverts commit 8a21ae07f9d4a568a1b4d3f4a558ac2ef009dfc1.

* Fix missing openmp symbols bug when staticlly link with ATen in Caffe2

When ATen is built as shared lib, the openmp dependency is added to the final so file with `-fopenmp` added to the link flags, instead of as `target_link_libraries` in cmake and so this dependency is not properly recorded in our build system. This diff is just a temporal solution, the proper one should be propergate USE_OPENMP from ATen to Caffe2 in cmake.